--- libs/apis/normal/index.js
@@ -1,7 +1,0 @@
-function ping(req, res) {
-  res.status(204).send();
-}
-
-module.exports = {
-  ping,
-};
