--- libs/db/index.js
@@ -1,5 +1,0 @@
-const db = require('./db');
-
-module.exports = {
-  ...db,
-};
