+++ libs/graphql/root/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  hello: () => 'Hello world yes!',
+};
