+++ package.json
@@ -15,9 +15,11 @@
     "express-graphql": "^0.12.0",
     "express-session": "^1.17.3",
     "graphql": "^16.6.0",
+    "graphql-compose": "^9.0.10",
+    "graphql-compose-mongoose": "^9.7.2",
     "http-errors": "~1.6.3",
     "jade": "~1.11.0",
-    "mongoose": "^6.7.5",
+    "mongoose": "^6.8.0",
     "morgan": "~1.9.1",
     "nodemon": "^2.0.20",
     "pm2": "^5.2.2",
