+++ routes/index.js
@@ -1,14 +1,18 @@
 /**
- * all of route in this file have the same base url '/';
+ * Route settings, match the app.use('/path')
+ *
+ * current: /
  */
+
 const express = require('express');
-const { ping } = require('../libs/apis/normal');
 
 const router = express.Router();
 
 /**
  * register router as //ping -> /ping
  */
-router.get('/ping', ping);
+router.get('/ping', (req, res) => {
+  res.sendStatus(204);
+});
 
 module.exports = router;
